admin/requirements.txt
Django==3.1.3
djangorestframework==3.12.2
mysqlclient==2.1.1
django-mysql==3.9
PyJWT==2.7.0
django-cors-headers==3.5.0